requirements.txt
transformers>=4.41.0
accelerate>=0.33.0
pyyaml>=6.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
tqdm>=4.66
